refactor(gateway): split accepted workspace lock script
